pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cmucal-backend"
version = "0.1.0"
description = "CMUCal backend scrapers and API"
requires-python = ">=3.10"

# -----------------------------
# Runtime dependencies ONLY
# -----------------------------
dependencies = [
    # Web / API
    "Flask>=3.1",
    "flask-cors",
    "Flask-JWT-Extended",
    "Flask-PyMongo",
    "gunicorn",

    # Database
    "SQLAlchemy>=2.0",
    "alembic",
    "psycopg2-binary",
    "pymongo",
    "supabase",

    # Auth / security
    "bcrypt",
    "Flask-Bcrypt",
    "PyJWT",
    "python-dotenv",

    # Scraping / parsing
    "requests",
    "urllib3",
    "beautifulsoup4",

    # Data / export
    "pandas",
    "numpy",
    "openpyxl",

    # Calendar / time
    "icalendar",
    "recurring-ical-events",
    "python-dateutil",
    "pytz",
    "tzlocal",
    "x-wr-timezone",

    # Google APIs
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",

    # LLM / Agents
    "langchain",
    "langchain-openai",
    "langgraph",
    "openai",
    "tiktoken",

    # Visualization (LangGraph)
    "graphviz",

    # Misc
    "inflect",
]

# -----------------------------
# Dev / test dependencies
# -----------------------------
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "coverage",
    "typeguard",
    "sqlacodegen",
]

# -----------------------------
# Package discovery
# -----------------------------
[tool.setuptools.packages.find]
where = ["."]
include = ["scraper*", "app*"]